updates on CourseCreate endpoint

+++ Udemy.Presentation/Controllers/CourseController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Udemy.Domain.Contracts;
+using Udemy.Domain.Enums;
 using Udemy.Domain.Models;
 using Udemy.Presentation.DTOs;
 using Udemy.Presentation.Helpers;
@@ -16,9 +17,8 @@ namespace Udemy.Presentation.Controllers
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
-        private readonly ApplicationDbContext _dbContext;
 
-        public CourseController(IUnitOfWork unitOfWork,IMapper mapper,ApplicationDbContext dbContext)
+        public CourseController(IUnitOfWork unitOfWork,IMapper mapper)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
@@ -38,10 +38,12 @@ namespace Udemy.Presentation.Controllers
         }
 
         [HttpPost("CreateCourse")]
-        public async Task<ActionResult<CourseDTO>> CreateCourse(CourseDTO course,IFormFile file)
+        public async Task<ActionResult<CourseDTO>> CreateCourse(CourseCreationDTO course,CourseLevel courseLevel)
         {
             if (!ModelState.IsValid)
                 return BadRequest();
+
+            course.Level = courseLevel.ToString();
 
             var newCourse = _mapper.Map<Course>(course);
             await _unitOfWork.CourseRepository.Add(newCourse);
@@ -55,8 +57,13 @@ namespace Udemy.Presentation.Controllers
         }
 
         [HttpDelete("DeleteCourse")]
-        public async Task<ActionResult> DeleteCourse(Course course)
+        public async Task<ActionResult> DeleteCourse(string Id)
         {
+            var course = await _unitOfWork.CourseRepository.GetByIdAsync(Id);
+
+            if (course is null)
+                return NotFound();
+
             _unitOfWork.CourseRepository.Delete(course);
             var Result = await _unitOfWork.CompleteAsync();
